Extract Talos' terraform upgrade to a separate group
The Talos provider does not have the same release cycle as the rest of
the Talos ecosystem.
This was generating a pull request to upgrade both Talos (the OS) and
the Talos provider at the same time.
We don't want to keep them in the same upgrade, so we are separating them.

Milestone: minor