Add a release script
Renovate is expecting a version reference (either a tag or branch name)
in its extends syntax.
I wanted "main" to be usable and always pointing to the latest version
of each preset.

Because of that, I needed a way to automatically update the version in
all JSON files when creating a new release.

Milestone: patch