Handle custom versions for Kubernetes and Talos
These two dependencies are very spepcial, and we had decided to manage
their version without any special comment.
This way of writting the Kubernetes version is needed in the
talconfig.yaml file.

I've taken the liberty to handle `talosVersion` as well, even if I'm not
sure if it is needed because I want to avoid possible confusion if it's
ever needed: the two pacakges are handle exactly the same!

Milestone: patch